sender/requirements.txt
backoff==2.2.1
click==8.1.3
Deprecated==1.2.13
googleapis-common-protos==1.57.0
grpcio==1.51.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.14.0
opentelemetry-distro==0.35b0
opentelemetry-exporter-otlp-proto-grpc==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
protobuf==3.20.3
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
